Return class, not instance, from OptionManager
The label, doc url, etc. values won't change for each class so they
might as well be class methods.

The `#form_fields` method relies on view helpers included in the class
so it is easier if it stays an instance method and we just make a class
method to wrap it.