Remove unnecessary dependency injection
In order to keep this test fast, we weren't loading the OptionSet model
(since it inherits from ActiveRecord::Base and we don't want to load
Rails). Instead we were injecting the model class into the `#submit` as
an optional argument.

This seems clunky since only our tests use this argument and it makes
the implementation slightly more abstract.

This commit stubs the OptionSet class directly in the context of the
test so it no longer needs to be passed in.