Extract default values into config
Previously, when you generated a vimrc and then reloaded the page, the
values in the form would reset to the defaults (since we were setting
the values in the form helpers).

This commit not only fixes that issue, but extracts all of the default
values into the OptionConfig class.

This commit also converts keys that get stored as JSON to strings.