Extract Compatible option to config file
This commit extracts the data/behavior specific to the "compatible"
option into a config file. This can now replace the partial entirely for
the compatible option form fields.

This pattern should be repeatable for the other options and future
options. The goal is to make adding additional options require editing
as few files as necessary.