Render "compatible" with Options::Type::Boolean
This moves the responsibility for rendering the "compatible" option from
the VimrcRenderer to the Options::Type::Boolean, which was already
responsible for rendering the form fields.

This will reduce duplication but does couple the form field rendering
concern with the rendering of the actual option. This is acceptable
since it seems likely for now that future options would act the
same in both areas.